Conditional export for 6.9 bug
In version 6.9 there is a bug which prevents users from creating conditional tasks where they ask a user in a task instead of an email. The work around is to copy and past an existing conditional. This playbook is a simple single task playbook which has the conditional example that can be copied.